Change module to lower-case, ready to build, will be a major version change
