--- pyLHF/src/LHF/DataGeneration/__init__.py
@@ -1,6 +1,0 @@
-from .dSphere import *
-from .torus import *
-from .embeddings import *
-from .projectivePlane import *
-from .dataGen import *
-from .objGen import *
